+++ renderer/core-html-renderer/src/main/groovy/com/ainrif/apiator/renderer/core/html/CoreHtmlRenderer.groovy
@@ -30,8 +30,6 @@ import static java.nio.file.StandardCopyOption.REPLACE_EXISTING
 import static java.util.Collections.emptyMap
 
 class CoreHtmlRenderer implements Renderer {
-    private static def logger = LoggerFactory.getLogger(CoreHtmlRenderer)
-
     protected String js
     protected String css
     protected String hbs